backend/requirements.txt
fastapi==0.115.0

uvicorn[standard]==0.30.6

pydantic==2.8.2

python-dotenv==1.0.1

cryptography>=42.0.0

loguru==0.7.2

sentry-sdk[fastapi]==2.13.0


httpx
requests
stripe

psycopg[binary]
prometheus-client==0.20.0
cryptography>=42.0.0
PyJWT==2.9.0
